+++ projects/swimlane/ngx-graph/src/stories/custom-templates/dagre-cluster.component.ts
@@ -1,9 +1,10 @@
-import { Component } from '@angular/core';
+import { Component, ChangeDetectionStrategy } from '@angular/core';
 import { NgxGraphModule } from '@swimlane/ngx-graph';
 
 @Component({
   selector: 'dagre-layout',
   imports: [NgxGraphModule],
+  changeDetection: ChangeDetectionStrategy.Eager,
   template: `
     <ngx-graph
       class="chart-container"