+++ projects/swimlane/ngx-graph/src/stories/demos/components/ngx-graph-msagl/ngx-graph-msagl.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, ChangeDetectionStrategy } from '@angular/core';
 import type { Layout } from '@swimlane/ngx-graph';
 import { NgxGraphModule } from '@swimlane/ngx-graph';
 import { MSAGLLayout } from './msaglLayout';
@@ -9,6 +9,7 @@ import * as shape from 'd3-shape';
   selector: 'ngx-graph-msagl',
   templateUrl: './ngx-graph-msagl.component.html',
   styleUrls: ['./ngx-graph-msagl.component.css'],
+  changeDetection: ChangeDetectionStrategy.Eager,
   imports: [NgxGraphModule]
 })
 export class NgxGraphMSAGLComponent {